+++ src/shapes/itext.class.ts
@@ -1,6 +1,10 @@
 // @ts-nocheck
 import { Canvas } from '../canvas/canvas_events';
-import { ObjectEvents, TPointerEventInfo } from '../EventTypeDefs';
+import {
+  ObjectEvents,
+  TPointerEvent,
+  TPointerEventInfo,
+} from '../EventTypeDefs';
 import { ITextClickBehaviorMixin } from '../mixins/itext_click_behavior.mixin';
 import {
   ctrlKeysMapDown,
@@ -13,9 +17,9 @@ import { classRegistry } from '../util/class_registry';
 
 export type ITextEvents = ObjectEvents & {
   'selection:changed': never;
-  changed: never;
+  changed: never | { index: number; action: string };
   tripleclick: TPointerEventInfo;
-  'editing:entered': never;
+  'editing:entered': never | { e: TPointerEvent };
   'editing:exited': never;
 };
 
@@ -346,12 +350,15 @@ export class IText extends ITextClickBehaviorMixin<ITextEvents> {
    * @param {number} index index from start
    * @param {boolean} [skipCaching]
    */
-  _getCursorBoundariesOffsets(index: number, skipCaching?: boolean) {
+  _getCursorBoundariesOffsets(
+    index: number,
+    skipCaching?: boolean
+  ): { left: number; top: number } {
     if (skipCaching) {
       return this.__getCursorBoundariesOffsets(index);
     }
     if (this.cursorOffsetCache && 'top' in this.cursorOffsetCache) {
-      return this.cursorOffsetCache;
+      return this.cursorOffsetCache as { left: number; top: number };
     }
     return (this.cursorOffsetCache = this.__getCursorBoundariesOffsets(index));
   }
@@ -476,10 +483,12 @@ export class IText extends ITextClickBehaviorMixin<ITextEvents> {
    * Renders drag start text selection
    */
   renderDragSourceEffect(this: AssertKeys<this, 'canvas'>) {
+    const dragStartSelection =
+      this.draggableTextDelegate.getDragStartSelection()!;
     this._renderSelection(
       this.canvas.contextTop,
-      this.__dragStartSelection,
-      this._getCursorBoundaries(this.__dragStartSelection.selectionStart, true)
+      dragStartSelection,
+      this._getCursorBoundaries(dragStartSelection.selectionStart, true)
     );
   }
 
@@ -618,6 +627,7 @@ export class IText extends ITextClickBehaviorMixin<ITextEvents> {
 
   dispose() {
     this._exitEditing();
+    this.draggableTextDelegate.dispose();
     super.dispose();
   }
 }